+++ lib/valid_email2/address.rb
@@ -45,7 +45,11 @@ module ValidEmail2
     end
 
     def disposable?
-      valid? && domain_is_in?(ValidEmail2.disposable_emails)
+      valid? &&
+        (
+          domain_is_in?(ValidEmail2.disposable_emails) ||
+          mx_server_is_in?(ValidEmail2.disposable_emails)
+        )
     end
 
     def whitelisted?
@@ -59,10 +63,7 @@ module ValidEmail2
     def valid_mx?
       return false unless valid?
 
-      Resolv::DNS.open do |dns|
-        return dns.getresources(address.domain, Resolv::DNS::Resource::IN::MX).size > 0 ||
-               dns.getresources(address.domain, Resolv::DNS::Resource::IN::A).size > 0
-      end
+      mx_servers.any?
     end
 
     private
@@ -74,10 +75,29 @@ module ValidEmail2
       }
     end
 
-    def address_contain_emoticons? email_str
-      return false if email_str.nil?
+    def mx_server_is_in?(domain_list)
+      mx_servers.any? { |mx_server|
+        return false unless mx_server.respond_to?(:exchange)
+        mx_server = mx_server.exchange.to_s
 
-      email_str.each_char.any? { |char| char.bytesize > 1 }
+        domain_list.any? { |domain|
+          mx_server.end_with?(domain) && mx_server =~ /\A(?:.+\.)*?#{domain}\z/
+        }
+      }
+    end
+
+    def address_contain_emoticons?(email)
+      return false if email.nil?
+
+      email.each_char.any? { |char| char.bytesize > 1 }
+    end
+
+    def mx_servers
+      @mx_servers ||= Resolv::DNS.open do |dns|
+        mx_servers = dns.getresources(address.domain, Resolv::DNS::Resource::IN::MX)
+        (mx_servers.any? && mx_servers) ||
+          dns.getresources(address.domain, Resolv::DNS::Resource::IN::A)
+      end
     end
   end
 end