+++ spec/valid_email2_spec.rb
@@ -34,95 +34,119 @@ describe ValidEmail2 do
   describe "basic validation" do
     subject(:user) { TestUser.new(email: "") }
 
-    it "should not be valid email is valid" do
+    it "is valid" do
       user = TestUser.new(email: "foo@bar123.com")
       expect(user.valid?).to be_truthy
     end
 
-    it "should be valid when email is empty" do
+    it "is invalid if email is empty" do
       expect(user.valid?).to be_truthy
     end
 
-    it "should not be valid when domain is missing" do
+    it "is invalid if domain is missing" do
       user = TestUser.new(email: "foo@.com")
       expect(user.valid?).to be_falsey
     end
 
-    it "should be invalid when domain includes invalid characters" do
+    it "is invalid if email contains invalid characters" do
       %w[+ _ !].each do |invalid_character|
         user = TestUser.new(email: "foo@google#{invalid_character}yahoo.com")
         expect(user.valid?).to be_falsey
       end
     end
 
-    it "should be invalid when email is malformed" do
+    it "is invalid if address is malformed" do
       user = TestUser.new(email: "foo@bar")
       expect(user.valid?).to be_falsey
     end
 
-    it "should be invalid if Mail::AddressListsParser raises exception" do
+    it "is invalid if Mail::AddressListsParser raises exception" do
       user = TestUser.new(email: "foo@gmail.com")
       expect(Mail::Address).to receive(:new).and_raise(Mail::Field::ParseError.new(nil, nil, nil))
       expect(user.valid?).to be_falsey
     end
 
-    it "shouldn't be valid if the domain constains consecutives dots" do
+    it "is invalid if the domain contains consecutive dots" do
       user = TestUser.new(email: "foo@bar..com")
       expect(user.valid?).to be_falsey
     end
+
+    it "is invalid if the domain contains emoticons" do
+      user = TestUser.new(email: "foo🙈@gmail.com")
+      expect(user.valid?).to be_falsy
+    end
   end
 
-  describe "disposable emails" do
-    it "should be valid when the domain is not in the list of disposable email providers" do
+  describe "with disposable validation" do
+    it "is valid if it's not a disposable email" do
       user = TestUserDisallowDisposable.new(email: "foo@gmail.com")
       expect(user.valid?).to be_truthy
     end
 
-    it "should be invalid when domain is in the list of disposable email providers" do
+    it "is invalid if it's a disposable email" do
       user = TestUserDisallowDisposable.new(email: "foo@#{ValidEmail2.disposable_emails.first}")
       expect(user.valid?).to be_falsey
     end
 
-    it "should be invalid when domain is a subdomain of a disposable domain" do
+    it "is invalid if the domain is a subdomain of a disposable domain" do
       user = TestUserDisallowDisposable.new(email: "foo@bar.#{ValidEmail2.disposable_emails.first}")
       expect(user.valid?).to be_falsey
     end
 
-    it "should allow example.com that is common in lists of disposable email providers" do
+    it "allows example.com" do
       user = TestUserDisallowDisposable.new(email: "foo@example.com")
       expect(user.valid?).to be_truthy
     end
 
+    context "with domain that is not disposable but it's mx server is disposable" do
+      let(:domain) { "sogetthis.com" }
+
+      around do |example|
+        ValidEmail2.disposable_emails.delete(domain)
+        example.run
+        ValidEmail2.disposable_emails << domain
+      end
+
+      it "is invalid" do
+        user = TestUserDisallowDisposable.new(email: "foo@sogetthis.com")
+        expect(user.valid?).to be_falsey
+      end
+    end
+
     describe "with whitelisted emails" do
-      it "should be invalid when the domain is in the list of disposable and there is no whitelist" do
-        user = TestUserDisallowDisposableWithWhitelist.new(email: "foo@#{ValidEmail2.disposable_emails.first}")
+      let(:whitelist_domain) { ValidEmail2.disposable_emails.first }
+      let(:whitelist_file_path) { "config/whitelisted_email_domains.yml" }
+
+      after do
+        FileUtils.rm(whitelist_file_path, force: true)
+      end
+
+      it "is invalid if the domain is disposable and not in the whitelist" do
+        user = TestUserDisallowDisposableWithWhitelist.new(email: "foo@#{whitelist_domain}")
         expect(user.valid?).to be_falsey
       end
 
-      it "should be valid when the domain is in the list of disposable but it is in the whitelist" do
-        whitelist_domain = ValidEmail2.disposable_emails.first
-        whitelist_file_path = "vendor/whitelist.yml"
-        File.open(whitelist_file_path, "w") { |f| f.write whitelist_domain.to_yaml }
+      it "is valid if the domain is disposable but in the whitelist" do
+        File.open(whitelist_file_path, "w") { |f| f.write [whitelist_domain].to_yaml }
         user = TestUserDisallowDisposableWithWhitelist.new(email: "foo@#{whitelist_domain}")
         expect(user.valid?).to be_falsey
-        File.delete(whitelist_file_path)
       end
     end
   end
 
-  describe "blacklisted emails" do
-    it "should be valid when email is not in the blacklist" do
+  describe "with blacklist validation" do
+    it "is valid if the domain is not blacklisted" do
       user = TestUserDisallowBlacklisted.new(email: "foo@gmail.com")
       expect(user.valid?).to be_truthy
     end
 
-    it "should be invalid when email is in the blacklist" do
+    it "is invalid if the domain is blacklisted" do
       user = TestUserDisallowBlacklisted.new(email: "foo@#{ValidEmail2.blacklist.first}")
       expect(user.valid?).to be_falsey
     end
   end
 
-  describe "mx lookup" do
+  describe "with mx validation" do
     it "should be valid if mx records are found" do
       user = TestUserMX.new(email: "foo@gmail.com")
       expect(user.valid?).to be_truthy
@@ -139,64 +163,35 @@ describe ValidEmail2 do
     end
   end
 
-  describe "emoticons emails" do
-    it "should be invalid if email contains emoticon" do
-      email = ValidEmail2::Address.new("foo🙈@gmail.com")
-      expect(email.valid?).to be_falsy
+  describe "with subaddress validation" do
+    it "is valid when address does not contain subaddress" do
+      user = TestUserSubaddressing.new(email: "foo@gmail.com")
+      expect(user.valid?).to be_truthy
+    end
+
+    it "is invalid when address cotains subaddress" do
+      user = TestUserSubaddressing.new(email: "foo+1@gmail.com")
+      expect(user.valid?).to be_falsey
     end
   end
 
-  describe "subaddressed emails" do
-    describe "::Address::DEFAULT_RECIPIENT_DELIMITER" do
-      it "should be recipient delimiter ('+')" do
-        expect(ValidEmail2::Address::DEFAULT_RECIPIENT_DELIMITER).to eq('+')
-      end
-    end
-
-    describe "::Address#subaddressed?" do
-      it "should be true when address local part contains a recipient delimiter ('+')" do
-        email = ValidEmail2::Address.new("foo+1@gmail.com")
-        expect(email.subaddressed?).to be_truthy
-      end
-
-      it "should be false when address local part contains a recipient delimiter ('+')" do
-        email = ValidEmail2::Address.new("foo@gmail.com")
-        expect(email.subaddressed?).to be_falsey
-      end
-    end
-
-    describe "user validation" do
-      context "subaddressing is allowed (default)" do
-        it "should be valid when address local part does not contain a recipient delimiter ('+')" do
-          user = TestUser.new(email: "foo@gmail.com")
-          expect(user.valid?).to be_truthy
-        end
-
-        it "should be valid when address local part contains a recipient delimiter ('+')" do
-          user = TestUser.new(email: "foo+1@gmail.com")
-          expect(user.valid?).to be_truthy
-        end
-      end
-
-      context "subaddressing is forbidden" do
-        it "should be valid when address local part does not contain a recipient delimiter ('+')" do
-          user = TestUserSubaddressing.new(email: "foo@gmail.com")
-          expect(user.valid?).to be_truthy
-        end
-
-        it "should be invalid when address local part contains a recipient delimiter ('+')" do
-          user = TestUserSubaddressing.new(email: "foo+1@gmail.com")
-          expect(user.valid?).to be_falsey
-        end
-      end
-    end
-  end
-
-  describe "custom error message" do
+  describe "with custom error message" do
     it "supports settings a custom error message" do
       user = TestUserMessage.new(email: "fakeemail")
       user.valid?
       expect(user.errors.full_messages).to include("Email custom message")
+    end
+  end
+
+  describe "#subaddressed?" do
+    it "should be true when address local part contains a recipient delimiter ('+')" do
+      email = ValidEmail2::Address.new("foo+1@gmail.com")
+      expect(email.subaddressed?).to be_truthy
+    end
+
+    it "should be false when address local part contains a recipient delimiter ('+')" do
+      email = ValidEmail2::Address.new("foo@gmail.com")
+      expect(email.subaddressed?).to be_falsey
     end
   end
 end